When hero is changed, all hero related purchases will be taken back, and user has to start his/her new hero from the beginning. Limited items will be available again, so, the other players can purchase those items again unless they are expired by time

+++ services/src/main/java/io/github/isuru/oasis/services/api/IProfileService.java
@@ -36,5 +36,5 @@ public interface IProfileService {
 
     boolean logoutUser(long userId, long ts) throws Exception;
 
-    boolean changeUserHero(long userId, long newHeroId) throws Exception;
+    boolean changeUserHero(long userId, int newHeroId) throws Exception;
 }
